+++ src/main/java/com/example/jsgamesbackendmain/Service/UserService.java
@@ -32,12 +32,6 @@ public class UserService {
     @Autowired
     private UserLoginBean userLoginBean;
 
-    @Autowired
-    private UserGetTop100Bean userGetTop100Bean;
-
-    @Autowired
-    private UserSetTop100Bean userSetTop100Bean;
-
     public UserGetResponseDTO getUser(Long userId) {
         return userGetBean.getUser(userId);
     }
@@ -56,16 +50,5 @@ public class UserService {
 
     public UserLoginResponseDTO loginUser(UserLoginRequestDTO userLoginRequestDTO){
         return userLoginBean.exec(userLoginRequestDTO);
-    }
-    public List<UserGetTop100ResponseDTO> getTop100User(Long page, Long size) {
-        List<UserTop100DAO> daoList = userGetTop100Bean.exec(page, size);
-        return daoList.stream().map(UserGetTop100ResponseDTO::of).collect(Collectors.toList());
-    }
-
-    public Map<String ,String > setTop100User() {
-        userSetTop100Bean.exec();
-        HashMap<String , String> map = new HashMap<>();
-        map.put("result", "success");
-        return map;
     }
 }
